Factored code better
Lots of the code has been rewritten so it is better factored, and more
useful. You can also dump UDP packets, although no checking is done that
the packet is a UDP packet, nor are the checksums calculated for those
packets and checks.

Some example code that needs porting has been added.

Another makefile target for testing has been added to listen for packets
as well.